Revert "feat: rename history components"
This reverts commit 746eb0826b986d1bb4aa2873cd576828761b6b5e.
